chore: 统一收纳check
