+++ .eslintrc.js
@@ -1,35 +1,6 @@
 // https://docs.expo.dev/guides/using-eslint/
-const forbiddenRule = require('./forbidden-rule');
-
-const absoluteColorPattern = String.raw`(^|[^A-Za-z0-9_-])#(?:[0-9a-fA-F]{3,4}|[0-9a-fA-F]{6}|[0-9a-fA-F]{8})(?![0-9a-fA-F])|(?:[rR][gG][bB][aA]?|[hH][sS][lL][aA]?)\s*\(`;
-const absoluteColorMessage =
-  '避免在业务 UI 中直接使用 hex/rgb/hsl 绝对颜色，请优先使用 Tailwind 语义 token 或集中业务色盘。';
-
-const noRawColorRule = [
-  'warn',
-  {
-    selector: `Literal[value=/${absoluteColorPattern}/]`,
-    message: absoluteColorMessage,
-  },
-  {
-    selector: `TemplateElement[value.raw=/${absoluteColorPattern}/]`,
-    message: absoluteColorMessage,
-  },
-];
-
-const adaptiveStyleOverride = {
-  files: [
-    'app/**/*.{js,jsx,ts,tsx}',
-    'components/**/*.{js,jsx,ts,tsx}',
-    'context/**/*.{js,jsx,ts,tsx}',
-    'hooks/**/*.{js,jsx,ts,tsx}',
-    'lib/**/*.{js,jsx,ts,tsx}',
-    'utils/**/*.{js,jsx,ts,tsx}',
-  ],
-  rules: {
-    'no-restricted-syntax': noRawColorRule,
-  },
-};
+const forbiddenRule = require('./tools/checks/forbidden-imports');
+const noRawColorOverride = require('./tools/checks/no-raw-color');
 
 /**
  * 深度比较两个值是否相等
@@ -115,7 +86,7 @@ module.exports = {
     ],
   },
   overrides: [
-    adaptiveStyleOverride,
+    noRawColorOverride,
     ...forbiddenRule
       .filter(item => item.allowIn?.length)
       .flatMap(item =>
