+++ babel.config.js
@@ -2,6 +2,6 @@ module.exports = function (api) {
   api.cache(true);
   return {
     presets: [['babel-preset-expo', { jsxImportSource: 'nativewind' }], 'nativewind/babel'],
-    plugins: ['./babel-plugin-forbidden-imports.js'],
+    plugins: ['./tools/checks/babel-plugin-forbidden-imports.js'],
   };
 };
