+++ tools/checks/babel-plugin-forbidden-imports.js
@@ -1,4 +1,4 @@
-const forbiddenRule = require('./forbidden-rule');
+const forbiddenRule = require('./forbidden-imports');
 const path = require('path');
 
 module.exports = function ({ types: t }) {
@@ -8,7 +8,6 @@ module.exports = function ({ types: t }) {
         const filename = state.file.opts.filename || '';
         const importSource = pathNode.node.source.value;
 
-        // 跳过 node_modules
         if (
           filename.includes(`${path.sep}node_modules${path.sep}`) ||
           filename.includes('/node_modules/') ||
@@ -17,25 +16,21 @@ module.exports = function ({ types: t }) {
           return;
         }
 
-        // 检查 ESLint 注释是否禁用本规则
         const leadingComments = pathNode.node.leadingComments || [];
         const hasESLintDisable = leadingComments.some(comment =>
           /eslint-disable(-next-line)?\s+no-restricted-imports/.test(comment.value),
         );
-        if (hasESLintDisable) return; // 放行
+        if (hasESLintDisable) return;
 
         for (const rule of forbiddenRule) {
-          // 模块名不匹配，跳过
           if (importSource !== rule.source) continue;
 
-          // 白名单路径匹配，跳过
           if (
             rule.allowIn?.some(allowPath => {
               const fileAbs = path.resolve(filename);
               const allowAbs = path.resolve(allowPath);
               const relative = path.relative(allowAbs, fileAbs);
 
-              // 文件名完全匹配，或在允许目录内
               return (
                 fileAbs === allowAbs || (relative !== '' && !relative.startsWith('..') && !path.isAbsolute(relative))
               );
@@ -44,7 +39,6 @@ module.exports = function ({ types: t }) {
             continue;
           }
 
-          // 检查是否导入了禁止的名称
           const banned = pathNode.node.specifiers.some(spec => {
             if (t.isImportSpecifier(spec)) {
               return rule.names.includes(spec.imported.name);
